Fix Shen'dralar Rep Values
Shen'dralar quests and turn-ins was bumped from 200 in Classic Era/Vanilla to 500 in Burning Crusade Classic. This has been incorrect for a while in terms of what BCs values should be tuned to. Also fix Foror's to represent the current name of Nostro's (even though it is just a comment).